// src/features/find-doctors/ui/FindDoctorsMessage.tsx
import { Badge, Flex, Paper, Stack, Text } from '@mantine/core';
import { DoctorSuggestion } from '@/entities/message/model/types';
import { getExperience } from '@/entities/message/lib/experience-formatter';
import InviteDoctorModal from './InviteDoctorModal';

interface FindDoctorsMessageProps {
    doctors: DoctorSuggestion[];
}

function FindDoctorsMessage({ doctors }: FindDoctorsMessageProps) {
    return (
        <Stack
            gap={5}
            style={{ alignSelf: 'flex-start' }}
            maw={{ base: 500, sm: 700 }}
        >
            <Text fw={800}>Подходящие специалисты</Text>
            <Stack gap={5}>
                {doctors.map((d) => (
                    <Paper
                        key={d.userId}
                        withBorder
                        bdrs={30}
                        px={{ base: 'md', sm: 'lg' }}
                        py={{ base: 'xs', sm: 'md' }}
                    >
                        <Flex justify={'space-between'}>
                            <Stack gap={7} flex={1}>
                                <Text fw={700}>{d.fullName}</Text>
                                <Text c="blue" size="sm">
                                    {getExperience(
                                        new Date(d.practiceStartDate),
                                    )}
                                </Text>

                                {d.specializations.length > 0 && (
                                    <Stack>
                                        {d.specializations.map((s, i) => (
                                            <Badge
                                                key={`${d.userId}-${i}`}
                                                variant="light"
                                                color="blue"
                                            >
                                                {s}
                                            </Badge>
                                        ))}
                                    </Stack>
                                )}

                                <Stack gap={10}>
                                    <Stack gap={2}>
                                        <Text size="sm" fw={600}>
                                            Адрес:
                                        </Text>
                                        <Text size="sm">{d.address}</Text>
                                    </Stack>

                                    <Stack gap={2}>
                                        <Text size="sm" fw={600}>
                                            Образование:
                                        </Text>
                                        <Text size="sm">{d.education}</Text>
                                    </Stack>
                                </Stack>

                                {d.description && (
                                    <Text size="md">{d.description}</Text>
                                )}
                            </Stack>

                            <InviteDoctorModal doctorId={parseInt(d.userId)} />
                        </Flex>
                    </Paper>
                ))}
            </Stack>
        </Stack>
    );
}

export default FindDoctorsMessage;
